pyproject.toml
[project]
name = "genbi-pythonpoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "blue>=0.9.1",
    "cffi>=1.17.1",
    "cryptography>=40.0.2",
    "langchain>=0.3.2",
    "langchain-community>=0.3.1",
    "matplotlib>=3.9.2",
    "numpy>=1.23",
    "oci>=2.129.1",
    "oracledb>=2.4.1",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "six>=1.16.0",
    "streamlit>=1.39.0",
    "psycopg2-binary>=2.9.9",
    "tabulate>=0.9.0",
    "langgraph>=0.2.34",
    "blinker>=1.8.2",
]
